requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
colorama==0.4.6
supabase==2.3.4
python-jose==3.3.0 